requirements.txt
svg.path==2.2
numpy==1.13.0
lxml==3.8.0
